refactor: move VideoInformation
